requirements.txt
PyGithub==2.8.1
kubernetes==34.1.0
openai==2.0.1
python-dotenv==1.1.1
requests==2.32.5
langchain-community==0.3.30
jinja2==3.1.4